# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Papers Please System API
  description: API для системы управления УПК (пограничные пункты)
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
paths:
...
  /api/v1/events:
    get:
      tags:
      - event-controller
      operationId: getAllEvents
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedResponseEventResponse"
    post:
      tags:
      - event-controller
      operationId: createEvent
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EventResponse"
  /api/v1/documents:
    get:
      tags:
      - document-controller
      operationId: getAllDocuments
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedResponseDocumentResponse"
    post:
      tags:
      - document-controller
      operationId: createDocument
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DocumentResponse"
...
  /api/v1/appeals:
    get:
      tags:
      - appeal-controller
      operationId: getAllAppeals
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedResponseAppealResponse"
    post:
      tags:
      - appeal-controller
      operationId: createAppeal
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppealRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AppealResponse"
  /api/v1/appeals/{id}/process:
    post:
      tags:
      - appeal-controller
      operationId: processAppeal
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: decision
        in: query
        required: true
        schema:
          type: string
          enum:
          - APPROVED
          - REJECTED
          - PUNISH_EXECUTOR
          - PUNISH_CHECKER
          - PUNISH_BOTH
      - name: decidedById
        in: query
        required: true
        schema:
          type: string
      - name: notes
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AppealResponse"
...
  /api/v1/events/{id}:
    get:
      tags:
      - event-controller
      operationId: getEventById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EventResponse"
    delete:
      tags:
      - event-controller
      operationId: deleteEvent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Unit"
    patch:
      tags:
      - event-controller
      operationId: partialUpdateEvent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventRequestPartial"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EventResponse"
  /api/v1/documents/{id}:
    get:
      tags:
      - document-controller
      operationId: getDocumentById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DocumentResponse"
    delete:
      tags:
      - document-controller
      operationId: deleteDocument
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Unit"
    patch:
      tags:
      - document-controller
      operationId: partialUpdateDocument
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentRequestPartial"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DocumentResponse"
  /api/v1/appeals/{id}:
    get:
      tags:
      - appeal-controller
      operationId: getAppealById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AppealResponse"
    delete:
      tags:
      - appeal-controller
      operationId: deleteAppeal
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Unit"
    patch:
      tags:
      - appeal-controller
      operationId: partialUpdateAppeal
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppealRequestPartial"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AppealResponse"
...
  /api/v1/events/by-shift/{shiftId}:
    get:
      tags:
      - event-controller
      operationId: getEventsByShift
      parameters:
      - name: shiftId
        in: path
        required: true
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedResponseEventResponse"
  /api/v1/documents/by-type/{type}:
    get:
      tags:
      - document-controller
      operationId: getDocumentsByType
      parameters:
      - name: type
        in: path
        required: true
        schema:
          type: string
          enum:
          - PASSPORT
          - VISA
          - CERTIFICATE
          - WORK_PERMIT
          - PHOTO
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedResponseDocumentResponse"
  /api/v1/appeals/by-ticket/{ticketId}:
    get:
      tags:
      - appeal-controller
      operationId: getAppealByTicket
      parameters:
      - name: ticketId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AppealResponse"
  /api/v1/appeals/by-filed-by/{userId}:
    get:
      tags:
      - appeal-controller
      operationId: getAppealsByFiledBy
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedResponseAppealResponse"
  /api/v1/appeals/by-decision/{decision}:
    get:
      tags:
      - appeal-controller
      operationId: getAppealsByDecision
      parameters:
      - name: decision
        in: path
        required: true
        schema:
          type: string
          enum:
          - APPROVED
          - REJECTED
          - PUNISH_EXECUTOR
          - PUNISH_CHECKER
          - PUNISH_BOTH
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedResponseAppealResponse"
components:
  schemas:
...
    EventRequest:
      required:
      - description
      - priority
      - shiftId
      type: object
      properties:
        shiftId:
          type: string
        description:
          maxLength: 500
          minLength: 5
          type: string
        priority:
          type: string
          enum:
          - LOW
          - NORMAL
          - HIGH
          - CRITICAL
    EventResponse:
      required:
      - description
      - id
      - priority
      - shiftId
      type: object
      properties:
        id:
          type: string
        shiftId:
          type: string
        description:
          type: string
        priority:
          type: string
          enum:
          - LOW
          - NORMAL
          - HIGH
          - CRITICAL
    DocumentRequest:
      required:
      - body
      - documentType
      - userId
      type: object
      properties:
        userId:
          type: string
        documentType:
          type: string
          enum:
          - PASSPORT
          - VISA
          - CERTIFICATE
          - WORK_PERMIT
          - PHOTO
        body:
          type: object
          additionalProperties:
            type: object
        validFrom:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
    DocumentResponse:
      required:
      - body
      - documentType
      - id
      - userId
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        documentType:
          type: string
          enum:
          - PASSPORT
          - VISA
          - CERTIFICATE
          - WORK_PERMIT
          - PHOTO
        body:
          type: object
          additionalProperties:
            type: object
        validFrom:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
...
    AppealRequest:
      required:
      - createdBy
      - status
      - ticketId
      type: object
      properties:
        ticketId:
          type: string
        createdBy:
          type: string
        status:
          type: string
          enum:
          - OPEN
          - IN_PROGRESS
          - NEED_INFO
          - CLOSED
          - REJECTED
        comment:
          type: string
    AppealResponse:
      required:
      - createdAt
      - createdBy
      - id
      - status
      - ticketId
      - verdict
      type: object
      properties:
        id:
          type: string
        ticketId:
          type: string
        createdBy:
          type: string
        status:
          type: string
          enum:
          - OPEN
          - IN_PROGRESS
          - NEED_INFO
          - CLOSED
          - REJECTED
        verdict:
          type: string
          enum:
          - APPROVED
          - REJECTED
          - PUNISH_EXECUTOR
          - PUNISH_CHECKER
          - PUNISH_BOTH
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
        checkedBy:
          type: string
        checkedAt:
          type: string
          format: date-time
...
    EventRequestPartial:
      type: object
      properties:
        shiftId:
          type: string
        description:
          maxLength: 500
          minLength: 5
          type: string
        priority:
          type: string
          enum:
          - LOW
          - NORMAL
          - HIGH
          - CRITICAL
    DocumentRequestPartial:
      type: object
      properties:
        documentType:
          type: string
          enum:
          - PASSPORT
          - VISA
          - CERTIFICATE
          - WORK_PERMIT
          - PHOTO
        body:
          type: object
          additionalProperties:
            type: object
        validFrom:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
    AppealRequestPartial:
      type: object
      properties:
        ticketId:
          type: string
        createdBy:
          type: string
        status:
          type: string
          enum:
          - OPEN
          - IN_PROGRESS
          - NEED_INFO
          - CLOSED
          - REJECTED
        verdict:
          type: string
          enum:
          - APPROVED
          - REJECTED
          - PUNISH_EXECUTOR
          - PUNISH_CHECKER
          - PUNISH_BOTH
        comment:
          type: string
        checkedBy:
          type: string
        checkedAt:
          type: string
          format: date-time
...
    PagedResponseEventResponse:
      required:
      - items
      - limit
      - offset
      - total
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/EventResponse"
        total:
          type: integer
          format: int64
        limit:
          type: integer
          format: int32
        offset:
          type: integer
          format: int32
    PagedResponseDocumentResponse:
      required:
      - items
      - limit
      - offset
      - total
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DocumentResponse"
        total:
          type: integer
          format: int64
        limit:
          type: integer
          format: int32
        offset:
          type: integer
          format: int32
    PagedResponseAppealResponse:
      required:
      - items
      - limit
      - offset
      - total
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AppealResponse"
        total:
          type: integer
          format: int64
        limit:
          type: integer
          format: int32
        offset:
          type: integer
          format: int32
    Unit:
      type: object
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT